moving files to a folder utils
