+++ models/contacts.json
@@ -68,5 +68,15 @@
     "name": "heuben kenry",
     "email": "tharetra.ukr@dictum.co.uk",
     "phone": "(728) 508-5793"
+  },
+  {
+    "name": "btrenre kenry",
+    "email": "tfdatewra.ukr@dictum.",
+    "phone": "(428) 598-9999"
+  },
+  {
+    "name": "Fennedy Mane",
+    "email": "tiewes.Cras@nonenimMauris.net",
+    "phone": "(342) 451-6666"
   }
 ]
